requirements.txt
aioxmpp==0.12.2
imutils==0.5.4
matplotlib==3.3.2
nidaqmx==0.5.7
numpy==1.19.5
opencv_python==4.5.1.48
pandas==1.1.3
Pillow==8.4.0
pypylon==1.7.2
PyYAML==6.0
scikit_image==0.18.2
scikit_learn==1.0.1
skimage==0.0
